services/batch/requirements.txt
SQLAlchemy
psycopg2-binary
scikit-learn
pandas
kafka-python
python-dotenv
joblib
